@@ -15,14 +15,14 @@
 /// A `LogHandler` is an implementation of a logging backend.
 ///
 /// This type is an implementation detail and should not normally be used, unless implementing your own logging backend.
-/// To use the SwiftLog API, please refer to the documentation of `Logger`.
+/// To use the SwiftLog API, please refer to the documentation of ``Logger``.
 ///
 /// # Implementation requirements
 ///
 /// To implement your own `LogHandler` you should respect a few requirements that are necessary so applications work
 /// as expected regardless of the selected `LogHandler` implementation.
 ///
-/// - The `LogHandler` must be a `struct`.
+/// - The ``LogHandler`` must be a `struct`.
 /// - The metadata and `logLevel` properties must be implemented so that setting them on a `Logger` does not affect
 ///   other `Logger`s.
 ///
@@ -113,7 +113,13 @@
 /// Please note that the above `LogHandler` will still pass the 'log level is a value' test above it iff the global log
 /// level has not been overridden. And most importantly it passes the requirement listed above: A change to the log
 /// level on one `Logger` should not affect the log level of another `Logger` variable.
-public protocol LogHandler {
+public protocol LogHandler: _SwiftLogSendableLogHandler {
+    /// The metadata provider this `LogHandler` will use when a log statement is about to be emitted.
+    ///
+    /// A ``Logger/MetadataProvider`` may add a constant set of metadata,
+    /// or use task-local values to pick up contextual metadata and add it to emitted logs.
+    var metadataProvider: Logger.MetadataProvider? { get set }
+
     /// This method is called when a `LogHandler` must emit a log message. There is no need for the `LogHandler` to
     /// check if the `level` is above or below the configured `logLevel` as `Logger` already performed this check and
     /// determined that a message should be logged.
@@ -164,6 +170,23 @@ public protocol LogHandler {
 }
 
 extension LogHandler {
+    /// Default implementation for `metadataProvider` which defaults to `nil`.
+    /// This default exists in order to facilitate source-compatible introduction of the `metadataProvider` protocol requirement.
+    public var metadataProvider: Logger.MetadataProvider? {
+        get {
+            nil
+        }
+        set {
+            #if DEBUG
+            if LoggingSystem.warnOnceLogHandlerNotSupportedMetadataProvider(Self.self) {
+                self.log(level: .warning, message: "Attempted to set metadataProvider on \(Self.self) that did not implement support for them. Please contact the log handler maintainer to implement metadata provider support.", metadata: nil, source: "Logging", file: #file, function: #function, line: #line)
+            }
+            #endif
+        }
+    }
+}
+
+extension LogHandler {
     @available(*, deprecated, message: "You should implement this method instead of using the default implementation")
     public func log(level: Logger.Level,
                     message: Logger.Message,
@@ -186,3 +209,11 @@ extension LogHandler {
                  line: line)
     }
 }
+
+// MARK: - Sendable support helpers
+
+#if compiler(>=5.6)
+@preconcurrency public protocol _SwiftLogSendableLogHandler: Sendable {}
+#else
+public protocol _SwiftLogSendableLogHandler {}
+#endif
